Projects/askthemenu/requirements.txt
pandas
numpy
sentence-transformers
scikit-learn
faiss-cpu
